requirements.txt
aiofiles==25.1.0
aiomysql==0.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncmy==0.2.11
bcrypt==5.0.0
black==26.1.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.2
cryptography==46.0.4
cssselect==1.3.0
cssutils==2.11.1
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
emails==0.6
et_xmlfile==2.0.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.20.3
flake8==7.3.0
greenlet==3.3.1
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.16
idna==3.11
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
librt==0.7.8
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
numpy==2.4.1
openpyxl==3.1.5
orjson==3.11.5
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
premailer==3.10.0
pyasn1==0.6.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyMySQL==1.1.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.0
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.51.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1; sys_platform != "win32"
virtualenv==20.36.1
watchfiles==1.1.1
websockets==16.0